Add basic features in the core engine
This commit adds few basic features in the core engine.

The first one is about the module class extending Sinatra::Base and
which gives some new helpers.

Those ones are mainly centered on the authentication feature with a
dedicated module.

Helpers for authentication are also present and can be include in an
application with the 'helpers' method from Sinatra.