Update the generator script (enhance, move and rename)
The generator script has moved from src/configure to bin/camiloo.

The script now is able to generate a basic final file tree in the
directory it is executed (call of Dir.pwd).

The content of the src directory is copied under <current dir>/lib/camiloo.

Then in order to work, the lib directory should be added in the load path
and 'require' statements should follow this naming convention.